adding loader inside core app
